InvoiceDTO added and invoice Item fetching issues resolved

+++ src/main/java/com/eternalcoders/pointedge/dto/InvoiceDTO.java
@@ -13,5 +13,5 @@ public class InvoiceDTO {
     private List<InvoiceItemDTO> items;
     private LocalDateTime date;
     private Double totalAmount;
-    private Integer loyaltyPoints;
+    private double loyaltyPoints;
 }
